requirements.txt
aioredis==2.0.1
asgiref==3.5.2
charset-normalizer==2.1.0
flask==2.2.1
httpx==0.23.0
hypercorn==0.13.2
loguru==0.6.0
pip-chill==1.0.1
pydantic==1.9.1
python-decouple==3.6
urllib3==1.26.11
